added GrahamScan

+++ src/optimizer/direct/StoppingCondition.cpp
@@ -1,12 +1,12 @@
 #include "StoppingCondition.h"
 
-StoppingCondition::StoppingCondition(int L, int M, int T) : L(L), M(M), mins(T) {
+StoppingCondition::StoppingCondition(unsigned long L, unsigned long M, int T) : L(L), M(M), mins(T) {
     if (T != 0) {
         time_eval = true;
     }
 }
 
-bool StoppingCondition::evaluate(int l, int m) {
+bool StoppingCondition::evaluate(unsigned long l, unsigned long m) {
     return (L == 0 || l < L) && (M == 0 || m < M) && (!time_eval || T > system_clock::now());
 }
